One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/pl-merch/pl-route-row/pl-route-row.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { PlRouteRowComponent } from './pl-route-row.component';

describe('Component: PlRouteRow', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [PlRouteRowComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([PlRouteRowComponent],
      (component: PlRouteRowComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(PlRouteRowComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(PlRouteRowComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-pl-route-row></app-pl-route-row>
  `,
  directives: [PlRouteRowComponent]
})
class PlRouteRowComponentTestController {
}

